% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{allowFunctions}
\alias{allowFunctions}
\title{Allow function calls without authorization}
\usage{
allowFunctions(funs = NULL)
}
\arguments{
\item{funs}{Vector of fucntion names to be added to the list. If is \code{NULL},
returns names of all currently allowed functions.}
}
\value{
Names of all currently allowed functions if \code{funs = NULL}.
}
\description{
This function adds fucntion names to the list of functions, which
can be called from the web page without manual confirmation in the R
session.
}
\seealso{
\code{\link{allowVariables}}, \code{\link{authorize}}, \code{\link{openPage}} (check argument
\code{allowedFunctions}), \code{\link{callFunction}}.
}
