% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{closePage}
\alias{closePage}
\title{Stop server}
\usage{
closePage()
}
\description{
Stops the server and closes all currently opened pages (if any). This function is a 
wrapper of \code{stopServer} method of class \code{\link{App}}.
}
\seealso{
\code{\link{openPage}}
}
