% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{closeSession}
\alias{closeSession}
\title{Close session}
\usage{
closeSession(sessionId = NULL)
}
\arguments{
\item{sessionId}{ID of a session to be closed. Can be a vector of IDs to close multiple session at once.
If \code{NULL}, all active sessions will be closed. (Without stopping the server, to stop server, please, use \link{closePage}.)}
}
\description{
Closes session with a given ID.
}
