% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{getPort}
\alias{getPort}
\title{Get number of the port on which the local server is running}
\usage{
getPort()
}
\description{
This function returns port number for the running server. By default, a random available port is used. One can also
set a port number as an argument of the \code{\link{openPage}} function. The port number can't be changed after the app
was initialized.This function
is a wrapper around method \code{getPort} of the class \code{\link{App}}.
}
\seealso{
\code{\link{openPage}}
}
