% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{getSessionIds}
\alias{getSessionIds}
\title{Get IDs of all active sessions}
\usage{
getSessionIds()
}
\value{
a vector of session IDs.
}
\description{
Returns IDs of all currently active sessions. An ID is a randomly generated combination of 6 letters and
numbers that is assigned to each session upon opening. This function is a wrapper around method \code{getSessionIds}
of class \code{\link{App}}.
}
