% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{removeMessage}
\alias{removeMessage}
\title{Removes a stored message}
\usage{
removeMessage(sessionId = NULL, messageId = NULL)
}
\arguments{
\item{sessionId}{ID of the session from where to remove a message. If there is only one active session, this argument
becomes optional.}

\item{messageId}{ID of the message to remove. If there is only one stored message, this agrument becomes optional.}
}
\description{
Removes a message with a given ID from the storage of a specified session. This functions is a wrapper around
method \code{removeMessage} of class \code{\link{Session}}.
}
\seealso{
\code{\link{authorize}}, \code{\link{getMessageIds}}.
}
