% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{sendHTML}
\alias{sendHTML}
\title{Send HTML to the server}
\usage{
sendHTML(html = "")
}
\arguments{
\item{html}{HTML code that will be added to the web page.}
}
\description{
Sends a piece of HTML code to the server and adds it at the end
or the \code{body} element.
}
\examples{
\donttest{sendHTML("Test...")
sendHTML("This is <b>bold</b>")
sendHTML("<table><tr><td>1</td><td>2</td></tr><tr><td>3</td><td>4</td></tr></table>")}
}
