% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{setSessionVariables}
\alias{setSessionVariables}
\title{Set session-specific variables}
\usage{
setSessionVariables(vars, sessionId = NULL)
}
\description{
Specifies variables that will be available (can be read or rewritten) only within a given session.
This is useful to safe state of the app for each client or for other personal settings. You can
also use it to limit user's access to data loaded in the current R session.
}
