% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{App}
\alias{App}
\title{App class}
\description{
Object of this class represents the entire jrc-based app. It stores all the active connections,
user-specific variables and all the global app settings. If you use exported wrapper functions, 
without specifying an app object, an instance of this class is stored in the package namespace.
You can always, retrieve it with the \code{\link{getPage}} function. However it is also possible
to run several apps simultaneously by creating new objects of this class and then managing them 
directly.
}
\section{Methods}{

\describe{
   \item{\code{getSession(sessionId = NULL)}}{
      Returns a session with a given ID or \code{NULL} if session with this ID doesn't exist. If \code{sessionId = NULL}
      and there is only one active session, returns it.
   }
   \item{\code{closeSession(session = NULL, inactive = NULL, old = NULL)}}{
      Closes websocket connection of the session and removes all the related data from the app. For more information on 
      the arguments, please, check \code{\link{closeSession}}
   }
   \item{\code{getSessionIds()}}{
       Returns IDs of all currently active sessions. 
   }
   \item{\code{startServer(port = NULL)}}{
       Starts a local server that listens to a given port. If \code{port = NULL}, picks up a random available port.
   }
   \item{\code{stopServer()}}{
       Closes all active sessions and stops a running server.
   }
   \item{\code{openPage(useViewer = TRUE, browser = NULL)}}{
      Opens a new web page either in browser, or in R Studio viewer. If \code{useViewer = FALSE}, but browser is not selected,
      a default installed browser is used. If browser is specified, \code{useViewer} is ignored. This method returns
      a new Session object, which should correspond to the page that has been just opened. However, if someone would start
      a new connection at the moment when this function is used, it may return a wrong session.
   }
   \item{\code{setEnvironment(envir)}}{
      Specifies an environment in which all the messages from the web pages will be evaluated. For more information,
      please, check \code{\link{setEnvironment}}.
   }
   \item{\code{allowFunctions(funs = NULL)}}{
      Adds function names to a list of allowed functions. These functions can be called from the web page without authorization
      from the R side. If \code{funs = NULL}, then returns a list of all currently allowed functions. For more information,
      please, check \code{\link{allowFunctions}}.
   }
   \item{\code{allowVariables(vars)}}{
      Adds variable names to  the list of allowed variables. These variables can be rewritten from the web page without 
      authorization from the R side. If \code{vars = NULL}, then returns a vector of all currently allowed variable names.
      For more information, please, check \code{\link{allowVariables}}.
   }
   \item{\code{startPage(path = NULL)}}{
      Sets path to the starting web page of the app. Path can be full or relative to the app's root directory. If 
      \code{path = NULL}, returns current path to the starting page.
   }
   \item{\code{rootDirectory(path = NULL)}}{
      Sets path to the root directory for the server. Any file, requested by the server, will be looked for in this directory.
      Can be a full path or a path relative to the current working directory. If \code{path = NULL}, returns path to the
      current root directory.
   }
   \item{\code{numberOfConnections(maxCon = NULL)}}{
      Sets maximum number of connections to be active simultaneously. If \code{maxCon = NULL}, returns current number of 
      allowed connections. Default value is \code{Inf}.
   }
   \item{\code{new(rootDirectory = NULL, startPage = NULL, onStart = NULL, 
   connectionNumber = Inf, allowedFunctions = c(), allowedVariables = c(), sessionVars = NULL)}}{
      Creates a new instance of class App. Check \code{\link{openPage}} documentaion for information about
      arguments.
   }
}
}

