% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{closeSession}
\alias{closeSession}
\title{Close one or several client sessions}
\usage{
closeSession(sessionId = NULL, inactive = NULL, old = NULL)
}
\arguments{
\item{sessionId}{IDs of the sessions to close. Can be a vector of multiple IDs.}

\item{inactive}{All sessions that were inactive (didn't receive any messages) for the
specified amount of time (in seconds) will be closed.}

\item{old}{All sessions that has been opened for at least specified amount of time (in seconds)
will be closed.}
}
\description{
Closes web socket connections for the selected client sessions and removes all the related
information from memory. This is a wrapper function around method \code{closeSession} of 
class \code{\link{App}}.
}
