% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{getPage}
\alias{getPage}
\title{Get opened page}
\usage{
getPage()
}
\value{
page-handling object if there is a currently opened jrc page, \code{NULL} otherwise.
}
\description{
Checks if there is a currently opened page. If so, returns an object with all
the information about the current session.
}
\details{
TO DO: Do we really need this function?
}
