% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{getSession}
\alias{getSession}
\title{Get a session}
\usage{
getSession(sessionId = NULL)
}
\arguments{
\item{sessionID}{ID of the session. If there is only one active session, this argument becomes optional.}
}
\value{
Object of class \code{\link{Session}}.
}
\description{
Returns \code{\link{Session}} by its ID. This function is a wrapper around method
\code{getSession} of class \code{\link{App}}.
}
