% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{getSessionIds}
\alias{getSessionIds}
\title{Get IDs of all active sessions}
\usage{
getSessionIds()
}
\value{
a \code{data.frame} with three columns: \code{id} - session ID, \code{startDate} - time and date
of the initialization of this session, \code{lastActive} - time and date of the last received message
from this web socket.
}
\description{
Returns IDs of all currently active sessions with date and time of their initialization
and the last received message.
}
