% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{getSessionVariable}
\alias{getSessionVariable}
\title{Get a variable from a client session environment}
\usage{
getSessionVariable(varName, sessionId = NULL)
}
\arguments{
\item{varName}{Name of the variable to search for. Must be a character.}

\item{sessionId}{ID of the session. If there is only one active session, this argument becomes optional.}
}
\value{
Requested variable
}
\description{
This function returns a variable, how it is seen from a session, e.g. for all the received function calls and
commands. It searches for the variable in the session environment and then, if variable is not found, checks enclosing 
frames on the environment, starting from the outer environment of the app (see \code{\link{setEnvironment}}). If the variable
doesn't exist, throws an error.
}
\details{
This function
is a wrapper around method \code{sessionVariables} of the class \code{\link{Session}}.
}
