% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{sendCommand}
\alias{sendCommand}
\title{Send a command to a web page}
\usage{
sendCommand(command, sessionId = NULL, wait = 0)
}
\arguments{
\item{command}{A line (or several lines separated by \code{\\n}) of JavaScript code. This code
will be immediately executed on the opened page. No R-side syntax check is performed.}

\item{sessionId}{An ID of the session to which the command should be sent. Can also be a vector of multiple session IDs.
If \code{NULL}, the command will be sent to all currently active sessions.}

\item{wait}{If \code{wait > 0}, after sending the message, R will wait for a reply for a given number of seconds. 
For this time (or until the reply is received), execution of other commands will be halted. Any incoming message 
from the session will be considered as a reply.}
}
\description{
\code{sendCommand} sends JavaScript code to the server and executes it on the specified
opened page. Use JavaScript function \code{jrc.sendCommand} to send R code from the web page
and execute it in the current R session. All commands send to R from the server will be executed
only after authorization in the currently running R session.
}
\details{
Each opened page gets its own environment, where all the commands are evaluated. Thus, any changes
made with the usual assignment operator \code{<-} will be limited to the page-specific environment. The changes
are still saved, but can be accessed only with \code{\link{getSessionVariable}} function. To make changes outside
of the page-specific environment use \code{<<-} instead. In JavaScript \code{windows.varibleName = "SomeValue"}
should be used instead of \code{varibleName = "SomeValue"}, if you want the variable to be accessible outside the
current \code{sendCommand} call.

This function is a wrapper around \code{sendCommand} method of class \code{\link{Session}}.
}
\examples{
 
\donttest{k <- 0
openPage()
sendCommand(paste0("button = document.createElement('input');",
              "button.type = 'button';",
              "button.addEventListener('click', function() {jrc.sendCommand('k <<- k + 1')});", 
              "button.value = '+1';",
              "document.body.appendChild(button);", collapse = "\n"))
closePage()}

}
\seealso{
\code{\link{authorize}}, \code{\link{sendData}}, \code{\link{sendHTML}}, \code{\link{callFunction}},
\code{\link{openPage}}.
}
