% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrc.R
\name{setEnvironment}
\alias{setEnvironment}
\title{Set Environment}
\usage{
setEnvironment(envir, sessionId = NULL)
}
\arguments{
\item{envir}{Environment where to evaluate the commands.}
}
\description{
Defines the environment, where the commands, received from the server, will be evaluated. By default,
\code{globalenv()} is used.
}
\examples{
setEnvironment(environment())

}
